Moved SessionStorageG into the use_cases.session_load package
